Second Addition Movement and Animations
After struggling to see what i was doing with a basic PC using unitys Capsule and Cubes as arms i decided to find some animations for movement and fighting. Using Blend tress to smoothen the animations. There is now some realistic movement and also still achieving the attitude of only moving straight in a ring. No rotation is allowed yet. i would like to keep it the way it currently is as some of the animations spin the player slightly.

Another day i will be making sure the Player can be forced back which will also work on NPCs. Forcing the player back also means adding a Head Hit animation and body hit animation. for these will be the two main attacks.

There will also be an illegal elbow which will be looked at much later and adding a referee to allow players to consider choice and a thinking process to see if using the elbow is a good idea.