Commit 3 Camera Functionality
I decided i wanted to see what i could do with the camera and focus the camera between 2 objects the PC and NPC so when fighting the camera had an effect to zoom in when the 2 fighters were close and zoom out when they were not.

Also allowing for so more combat options and thinking about how i will be using my stamina system to effect the amount of force applied in every type of punch.