Big Changes
So after a lot of struggle i decided to change the IDE physics to a Rigidbody instead of using a character controller. I wanted to use the effect of a knockback pushing the player or NPC back with force however after many approaches to the issue i decided to change the physics and used the Rigidbody AddForce function to acheive the effect i wanted now the player is knocked when the player enters the appropiate trigger. This will now be intergrated around the player and NPCs fist. It also leaves room later for how much force is applied depending on the stamina value