Fixing the root motion
In this commit my root motion was causing it so the player could'nt move.

However i needed the root motion on the animator active because when trying to lead jab or body punch the models body would rotate in the wrong direction.

The commit has fixed this by baking the position of the animation clip rotation and now the animation works fine.

Next time i will be starting to create it so the value of the stamina effects the value of the force and allow the player and NPC can regenerate stamina when not moving so creating a state.

I will then start to create  a AI state machine which allows the player to monitor its states as well as increasing in difficulty when player fights in further rounds.