Making the referee move
This commit i made a simple cube move between 2 fighters. Because if i am to continue with making my Player fighter logic and implementing the illegal elbow states

I needed the referee to move between the 2 fighters to give the immersion of a referee monitoring a fight. Also editing some comments in my base class and GM

I also needed to make the referee as if the player slips out an illegal elbow the referee needs to decide if its seen or not seen. Which depending on the state and choice will then later effect the crowd and also the game state and risk between the 2 fighters.