Back and forwards action
In this commit a few important factors happened to get the dice rolloing with the pattern of the games structure.

Now there is a standard fighter who will go towards the player and throw a jab at them. the Player and NPC have set up colliders which detect relevant triggers on the 2 fighters hands which will be used to pass data from an animation effect to being punched stamina loss and other things to come. Both fighters can push eachother back and have the toe to toe action i looked to acheive.

The AI right now is however basic and needs more work to run smooth. the animations need adjusting as there are some small errors with clips clashing.

The next aim for me is to allow the NPC Fighter to choose what attack it will decide to use out of 3 having some effects when stamina levels reach a low value. Fixing any on going issues ith animations and balancing the timing and NPC can fight so i can create a balanced stat changer when the PC win.