Man in the crowed
In this commit i made the maths behind the flying bottle object to go from the angry crowed members hand to the position of the referee or the offender being 1 of the 2 fighters.

This mechanic is implemented in as there will be that one deticated fan in most sports where they wont see eye to eye with the referees choices. So in the game the ref can see or not see an illegal elbow to the face

if the ref sees the offence then the fan is fine no issues are found. However, the moment that referee goes blinde the fan will launch an object into the ring.

The object destroies itself if hitting one of its choosen objects.
However soon it'll daze the 2 fighters or the ref leading to faster gameplay the ref being dazed means illegal elbows can be dealt out and there is no consequences but if a ighter is dazed will then the opponent can just blitz the opponent until they're back in the fighting state.