Wrapping Up
In this commit i added some UI to feedback to the player being 2 Slider Health bars for both fighters.

I also allowed the NPC fighter to respawn when defeated and spawn a random fighter depending on the int that gets generated by our switch statement in the NPC derived class

I also allowed the NPC to stand away from the player when the player dies and making a game over screen appear when the player dies also a main menu screen to start and quit the game

I also started to allow the GameManager to spawn the objects and balance the UI except the health bars.

There is also a round indicator to show players how many NPCs they have defeated

I also fixed an issue with the NPCs combat where they would attack air even when the PC was no where near them.