Wrapped Up
In this commit i did the final edits of the game.
Changing and balancing stats for the NPC fighters, making a controls menu so players know how to play.
Also fixing a referee issue where when the player would illegal elbow the timers would not reset so when the right mouse button was pressed again the ref would glitch back and forwards.
I also made it so that the players health and stamina would reset when beating an opponent.